1. For the time being we need a pointer to the corresponding    xBin from an xPage 2. Fixes typo for defining __XMALLOC_DEBUG in debug version 3. xFindInSortedList takes xSpecBin, not xBin

+++ tests/t1.c
@@ -9,7 +9,7 @@
 
 #include <stdio.h>
 #include <xmalloc-config.h>
-#include "src/xmalloc.h"
+#include "../src/xmalloc.h"
 
 int main()
 {
@@ -19,7 +19,7 @@ int main()
     xBin b=xGetSpecBin(i);
     void *p=xMalloc(i);
     if ((i>b->sizeInWords*4)
-    || (b->sizeInWords*4!=xSizeOfAddr(p)))
+    || (b->sizeInWords*4 != (long) xSizeOfAddr(p)))
       printf(" %d (%d vs %d)\n",i,((int) b->sizeInWords)*4,(int) xSizeOfAddr(p));
     xFree(p);
   }
